perf(dsv4): load compressor wkv/wgate as bf16 in weight loader
The fused wkv+wgate GEMM in CompressorFP8 runs in bf16, but the loader
was forcing wkv/wgate to fp32 (legacy fp32-GEMM convention), causing a
bf16 → fp32 → bf16 round-trip at module init. Match the ckpt's native
bf16 dtype directly. Keep the .to(bfloat16) at the bind site as a
defensive no-op for non-default loaders.

+++ rtp_llm/models/deepseek_v4.py
@@ -187,20 +187,19 @@ class DeepSeekV4Weight(DeepSeekV2Weight):
         norm_name = W.v4_indexer_compressor_norm if inner else W.v4_compressor_norm
         ape_name = W.v4_indexer_compressor_ape if inner else W.v4_compressor_ape
         return [
-            # wkv / wgate are consumed by the compressor's fp32 ``F.linear``
-            # math; load straight to fp32 here so the bind site doesn't have
-            # to cast (one less bf16 -> fp32 transient at module init).
+            # wkv / wgate are consumed by the compressor's bf16 fused GEMM;
+            # keep them in the ckpt's native bf16 — no cast at the bind site.
             AtomicWeight(
                 wkv_name,
                 [CkptWeightInfo(f"layers.{{i}}.{ckpt_prefix}.wkv.weight", identity)],
                 identity,
-                data_type=torch.float32,
+                data_type=torch.bfloat16,
             ),
             AtomicWeight(
                 wgate_name,
                 [CkptWeightInfo(f"layers.{{i}}.{ckpt_prefix}.wgate.weight", identity)],
                 identity,
-                data_type=torch.float32,
+                data_type=torch.bfloat16,
             ),
             AtomicWeight(
                 norm_name,
